Refactor navigation ID type from int to short
Updated INavigationService to use short for navigationId.
Removed default constructor in EndpointWithCsvBase and
EndpointWithCsvTotalsBase, adding a new constructor for
short navigationId. Modified UpdateNavigationStatusEndpoint
to cast navigationId from int to short. Added using
directive for navigation entities in several endpoints.
Updated endpoint constructors to use Navigation.Constants.
Adjusted NavigationServiceTests to correctly pass
cancellationToken. These changes enhance consistency and
reduce potential issues with identifier size.

+++ src/services/src/Demoulas.ProfitSharing.Common/Interfaces/Navigations/INavigationService.cs
@@ -4,7 +4,7 @@ namespace Demoulas.ProfitSharing.Common.Interfaces.Navigations;
 public interface INavigationService
 {
     Task<List<NavigationDto>> GetNavigation(CancellationToken cancellationToken);
-    NavigationDto GetNavigation(int navigationId);
+    NavigationDto GetNavigation(short navigationId);
     Task<List<NavigationStatusDto>> GetNavigationStatus(CancellationToken cancellationToken);
-    Task<bool> UpdateNavigation(int navigationId, byte statusId, CancellationToken cancellationToken);
+    Task<bool> UpdateNavigation(short navigationId, byte statusId, CancellationToken cancellationToken);
 }
